+++ tabulous/_qt/_table/_base/_line_edit.py
@@ -1,0 +1,75 @@
+from __future__ import annotations
+from typing import TYPE_CHECKING
+from qtpy import QtWidgets as QtW, QtCore, QtGui
+from qtpy.QtCore import Qt
+from ..._keymap import QtKeys
+
+
+if TYPE_CHECKING:
+    from ._table_base import QMutableTable
+    from ._enhanced_table import _QTableViewEnhanced
+
+
+class QTableLineEdit(QtW.QLineEdit):
+    """LineEdit widget with dtype checker and custom defocusing."""
+
+    def __init__(
+        self,
+        parent: QtCore.QObject | None = None,
+        table: QMutableTable | None = None,
+        pos: tuple[int, int] = (0, 0),
+    ):
+        super().__init__(parent)
+        self._table = table
+        self._pos = pos
+        self.textChanged.connect(self.onTextChanged)
+
+    def parentTableView(self) -> _QTableViewEnhanced:
+        return self.parent().parent()
+
+    def isTextValid(self, r: int, c: int, text: str) -> bool:
+        """True if text is valid for this cell."""
+        raise NotImplementedError()
+
+    def onTextChanged(self, text: str) -> None:
+        """Change text color to red if invalid."""
+        palette = QtGui.QPalette()
+        if self.isTextValid(self._pos[0], self._pos[1], text):
+            col = Qt.GlobalColor.black
+        else:
+            col = Qt.GlobalColor.red
+
+        palette.setColor(QtGui.QPalette.ColorRole.Text, col)
+        self.setPalette(palette)
+        return None
+
+    def keyPressEvent(self, event: QtGui.QKeyEvent) -> None:
+        """Handle key press events."""
+        keys = QtKeys(event)
+        pos = self.cursorPosition()
+        nchar = len(self.text())
+        r, c = self._pos
+        if keys.is_moving():
+            if pos == 0 and keys == "Left" and c > 0:
+                self.parentTableView().setFocus()
+                self._table._qtable_view._selection_model.move_to(r, c - 1)
+                return
+            elif (
+                pos == nchar
+                and keys == "Right"
+                and c < self._table.model().columnCount() - 1
+                and self.selectedText() == ""
+            ):
+                self.parentTableView().setFocus()
+                self._table._qtable_view._selection_model.move_to(r, c + 1)
+                return
+            elif keys == "Up" and r > 0:
+                self.parentTableView().setFocus()
+                self._table._qtable_view._selection_model.move_to(r - 1, c)
+                return
+            elif keys == "Down" and r < self._table.model().rowCount() - 1:
+                self.parentTableView().setFocus()
+                self._table._qtable_view._selection_model.move_to(r + 1, c)
+                return
+
+        return super().keyPressEvent(event)
